Enable compiler flag -Wsign-compare
This flag causes Clang to emit a warning when it encounters code
that is comparing signed vs unsigned integers.

The reason for enabling this is that in the Fuchsia build this is
both enabled and considered an error not just a warning. So
all such transgressions must be fixed in order to build Cobalt for
Fuchsia.

The fixes are all really mundane:
- replace int with size_t for the loop variable in a for loop
- Replace integer literals such as "0" with the unsigned version
such as "0U".

We also unset this flag (and a couple of other flags) prior to
compiling the third_party directories since we are not going to
edit the code in those directories in the Cobalt repo there is
no point emitting compiler warnings for them.

+++ algorithms/forculus/field_element.cc
@@ -215,7 +215,7 @@ const byte* FieldElement::KeyBytes() const {
 
 std::ostream& operator<<(std::ostream& os, const FieldElement& el) {
   const byte* data = el.KeyBytes();
-  for (int i = 0; i < FieldElement::kDataSize; i++) {
+  for (size_t i = 0; i < FieldElement::kDataSize; i++) {
      os << std::hex << std::setfill('0') << std::setw(2) <<
          static_cast<uint32_t>(*data) << " ";
      data++;
