+++ algorithms/forculus/forculus_analyzer_test.cc
@@ -101,8 +101,8 @@ TEST(ForculusAnalyzerTest, NoErrors) {
   // 22 * 8 observations of plaintext4 on day 3.
   AddObservations(&forculus_analyzer, 3, DAY, plaintext4, kThreshold + 2, 8);
 
-  EXPECT_EQ(0, forculus_analyzer.observation_errors());
-  static const int kExpectedNumObservations =
+  EXPECT_EQ(0u, forculus_analyzer.observation_errors());
+  static const size_t kExpectedNumObservations =
       kThreshold * 5 + kThreshold * 5 +
       (kThreshold + 1) * 6 + (kThreshold - 1) * 6 +
       (kThreshold - 1) * 7 + (kThreshold - 1) * 7 +
@@ -111,19 +111,19 @@ TEST(ForculusAnalyzerTest, NoErrors) {
   auto results = forculus_analyzer.TakeResults();
 
   // We should have decrypted plaintexts 1, 2 and 4.
-  EXPECT_EQ(3, results.size());
+  EXPECT_EQ(3u, results.size());
 
   // Check plaintext1
   EXPECT_EQ(kThreshold * 5 + kThreshold * 5, results[plaintext1]->total_count);
-  EXPECT_EQ(2, results[plaintext1]->num_epochs);
+  EXPECT_EQ(2u, results[plaintext1]->num_epochs);
 
   // Check plaintext2
   EXPECT_EQ((kThreshold + 1) * 6, results[plaintext2]->total_count);
-  EXPECT_EQ(1, results[plaintext2]->num_epochs);
+  EXPECT_EQ(1u, results[plaintext2]->num_epochs);
 
   // Check plaintext4
   EXPECT_EQ((kThreshold + 2) * 8, results[plaintext4]->total_count);
-  EXPECT_EQ(1, results[plaintext4]->num_epochs);
+  EXPECT_EQ(1u, results[plaintext4]->num_epochs);
 
   // Plaintext3 should not be decrypted.
   EXPECT_EQ(nullptr, results[plaintext3]);
@@ -149,7 +149,7 @@ TEST(ForculusAnalyzerTest, TestEpochTypes) {
   // Since day 0 and day 1 are different epochs we should not have decrypted the
   // plaintext.
   auto results = forculus_analyzer->TakeResults();
-  EXPECT_EQ(0, results.size());
+  EXPECT_EQ(0u, results.size());
 
   // Next we test with a WEEK epoch
   forculus_config.set_epoch_type(WEEK);
@@ -164,7 +164,7 @@ TEST(ForculusAnalyzerTest, TestEpochTypes) {
   // Since day 0 and day 1 are in the same epoch we should have decrypted the
   // plaintext.
   results = forculus_analyzer->TakeResults();
-  EXPECT_EQ(1, results.size());
+  EXPECT_EQ(1u, results.size());
 
   // Next we test with a WEEK epoch but two days in different weeks.
   forculus_analyzer.reset(new ForculusAnalyzer(forculus_config));
@@ -178,7 +178,7 @@ TEST(ForculusAnalyzerTest, TestEpochTypes) {
   // Since day 0 and day 7 are different epochs we should not have decrypted the
   // plaintext.
   results = forculus_analyzer->TakeResults();
-  EXPECT_EQ(0, results.size());
+  EXPECT_EQ(0u, results.size());
 
   // Next we test with a MONTH epoch
   forculus_config.set_epoch_type(MONTH);
@@ -193,7 +193,7 @@ TEST(ForculusAnalyzerTest, TestEpochTypes) {
   // Since day 0 and day 7 are in the same epoch we should have decrypted the
   // plaintext.
   results = forculus_analyzer->TakeResults();
-  EXPECT_EQ(1, results.size());
+  EXPECT_EQ(1u, results.size());
 
   // Finally we test with a MONTH epoch but two days in different months.
   forculus_config.set_epoch_type(MONTH);
@@ -208,7 +208,7 @@ TEST(ForculusAnalyzerTest, TestEpochTypes) {
   // Since day 0 and day 31 are in different epochs we should not have decrypted
   // the  plaintext.
   results = forculus_analyzer->TakeResults();
-  EXPECT_EQ(0, results.size());
+  EXPECT_EQ(0u, results.size());
 }
 
 // Tests the use of a ForculusAnalyzer when fed observations with errors.
@@ -224,54 +224,54 @@ TEST(ForculusAnalyzerTest, WithErrors) {
 
   // Add an observation.
   EXPECT_TRUE(forculus_analyzer.AddObservation(0, obs));
-  EXPECT_EQ(1, forculus_analyzer.num_observations());
-  EXPECT_EQ(0, forculus_analyzer.observation_errors());
+  EXPECT_EQ(1u, forculus_analyzer.num_observations());
+  EXPECT_EQ(0u, forculus_analyzer.observation_errors());
 
   // Now add another observation with the same ciphertext and x-value
   // but a different y-value. This causes an error.
   obs.set_point_y("2 y fake");
   EXPECT_FALSE(forculus_analyzer.AddObservation(0, obs));
-  EXPECT_EQ(1, forculus_analyzer.num_observations());
-  EXPECT_EQ(1, forculus_analyzer.observation_errors());
+  EXPECT_EQ(1u, forculus_analyzer.num_observations());
+  EXPECT_EQ(1u, forculus_analyzer.observation_errors());
 
   // The whole ciphertext is considered corupt now so even changing to
   // a differnt x value still yields an error.
   obs.set_point_x("2 x fake");
   EXPECT_FALSE(forculus_analyzer.AddObservation(0, obs));
-  EXPECT_EQ(1, forculus_analyzer.num_observations());
-  EXPECT_EQ(2, forculus_analyzer.observation_errors());
+  EXPECT_EQ(1u, forculus_analyzer.num_observations());
+  EXPECT_EQ(2u, forculus_analyzer.observation_errors());
 
   // Adding an observation for a different epoch succeeds.
   obs.set_point_x("1 x fake");
   obs.set_point_y("1 y fake");
   EXPECT_TRUE(forculus_analyzer.AddObservation(1, obs));
-  EXPECT_EQ(2, forculus_analyzer.num_observations());
-  EXPECT_EQ(2, forculus_analyzer.observation_errors());
+  EXPECT_EQ(2u, forculus_analyzer.num_observations());
+  EXPECT_EQ(2u, forculus_analyzer.observation_errors());
 
   // Adding a second obseration for epoch 1 also succeeds.
   obs.set_point_x("2 x fake");
   obs.set_point_y("2 y fake");
   EXPECT_TRUE(forculus_analyzer.AddObservation(1, obs));
-  EXPECT_EQ(3, forculus_analyzer.num_observations());
-  EXPECT_EQ(2, forculus_analyzer.observation_errors());
+  EXPECT_EQ(3u, forculus_analyzer.num_observations());
+  EXPECT_EQ(2u, forculus_analyzer.observation_errors());
 
   // Adding a third obseration for epoch 1 invokes a decryption which fails.
   obs.set_point_x("3 x fake");
   obs.set_point_y("3 y fake");
   EXPECT_FALSE(forculus_analyzer.AddObservation(1, obs));
-  EXPECT_EQ(3, forculus_analyzer.num_observations());
-  EXPECT_EQ(3, forculus_analyzer.observation_errors());
+  EXPECT_EQ(3u, forculus_analyzer.num_observations());
+  EXPECT_EQ(3u, forculus_analyzer.observation_errors());
 
   // Adding a fourth obseration for epoch 1 also fails.
   obs.set_point_x("4 x fake");
   obs.set_point_y("4 y fake");
   EXPECT_FALSE(forculus_analyzer.AddObservation(1, obs));
-  EXPECT_EQ(3, forculus_analyzer.num_observations());
-  EXPECT_EQ(4, forculus_analyzer.observation_errors());
+  EXPECT_EQ(3u, forculus_analyzer.num_observations());
+  EXPECT_EQ(4u, forculus_analyzer.observation_errors());
 
   // There should be no results.
   auto results = forculus_analyzer.TakeResults();
-  EXPECT_EQ(0, results.size());
+  EXPECT_EQ(0u, results.size());
 }
 
 }  // namespace forculus
